requirements.txt
streamlit>=1.20.0
youtube-transcript-api>=0.6.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4.post2
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0
